+++ app/src/main/java/com/helloandroid/list_characters/ListCharactersContract.kt
@@ -18,4 +18,4 @@ interface ListCharactersContract {
     }
 }
 
-class CharacterItem(val id: Int, val name: String, val hp: Int, val skills: List<Pair<String, Int>>, val things: List<Pair<String, Int>>, var index: Int = -1)
+class CharacterItem(val id: Long, val name: String, val hp: Int, val skills: List<Pair<String, Int>>, val things: List<Pair<String, Int>>, var index: Int = -1)
